Move TestAddressRepository from ach to ui-core
COAND-368
